+++ src/main/java/dat3/voximovies/repository/MovieRepository.java
@@ -3,9 +3,9 @@ package dat3.voximovies.repository;
 import dat3.voximovies.entity.Movie;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MovieRepository extends JpaRepository<Movie, Integer> {
+public interface MovieRepository extends JpaRepository<Movie, Long> {
 
-  Movie findMovieById(int id);
+  Movie findMovieById(long id);
 
   boolean existsByName(String name);
